refactor code and use letter component
